fix repositories

+++ src/main/java/com/we/repository/SubCategoryRepository.java
@@ -3,5 +3,5 @@ package com.we.repository;
 import com.we.model.SubCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubCategoryRepository extends JpaRepository<SubCategory,Integer> {
+public interface SubCategoryRepository extends JpaRepository<SubCategory,Long> {
 }
